legacy-service/requirements.txt
tortoise-orm
asyncpg
pandas